pyproject.toml
[project]
name = "enka-py"
requires-python = ">=3.10"

[tool.poetry]
authors = ["seriaati <seria.ati@gmail.com>"]
description = "An async API wrapper for enka.network written in Python."
license = "GPL-3.0-only"
name = "enka-py"
packages = [{include = "enka"}]
readme = "README.md"
repository = "https://github.com/seriaati/enka-py"
version = "1.7.1"

[tool.poetry.dependencies]
aiofiles = "^23.2.1"
aiohttp-client-cache = {extras = ["sqlite"], version = "^0.11.0"}
orjson = "^3.9.10"
pydantic = "^2.5.3"
python = "^3.10"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.3"
ruff = "^0.4.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = [
  "E501",
  "ANN401",
  "ANN101",
  "ANN102",
  "PLR0913",
  "PLR2004",
  "PLR0917",
  "PLR6301",
  "ANN002",
  "ERA001",
  "RUF100",
  "D100",
  "D104",
  "D105",
  "D107",
]
preview = true
select = [
  "E",
  "W",
  "C90",
  "F",
  "UP",
  "B",
  "SIM",
  "I",
  "N",
  "TCH",
  "ANN",
  "ASYNC",
  "A",
  "C4",
  "EM",
  "FA",
  "ICN",
  "G",
  "PIE",
  "T20",
  "ARG",
  "ERA",
  "LOG",
  "PL",
  "TRY",
  "D",
  "RUF",
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F403", "F401"]
"enka/gi.py" = ["F403", "F401"]
"enka/hsr.py" = ["F403", "F401"]
"enka/models/*.py" = ["ALL"]
"example.py" = ["T201", "ERA001"]
"test.py" = ["D", "T201"]
"tests/*.py" = ["D"]

[tool.ruff.lint.flake8-type-checking]
quote-annotations = true

[tool.ruff.lint.pydocstyle]
convention = "google"
